requirement.txt
torch
pillow
tensorflow==2.2.0
matplotlib
torchvision
PyYAML
uvicorn
opencv-python
numpy
starlette
google
